Copy bytes when cloning encoder
Rather than using slice tricks, grab an encoder from the pool and copy
into it. This potentially re-uses an already-allocated backing array.
Since we were going to copy the underlying bytes as soon as a field was
added, it's better to eat the copying cost immediately and use
a pre-warmed buffer.