Add comparison benchmarks
Add benchmarks that compare zap to logrus, log15, go-kit, and the
standard lib. All the benchmarks (except the standard lib, obviously)
test three scenarios: logging a message with 10 fields added at the call
site, logging a message with no fields on a logger that already has 10
fields of context, and logging a message with no fields on a fresh
logger.

Spoiler: zap crushes the competition.