Add a field factory for stacktraces
Add a field factory that wraps `runtime.Stack`, so that users don't need
to pick a byte size for their stacktraces. The wrapper attempts to
re-use a buffer from the pool, but allocates if it's necessary to
capture the whole stacktrace.

Note that the Stack() function is *really* expensive - casting the trace
to a string costs an allocation, but just taking a stacktrace is ~10
microseconds (roughly the same amount of time as logging 100
non-stacktrace fields).