Add a Hooks option
Extend the logger to support a variable number of pre-log hooks. The
hooks can alter the in-flight message and/or add to the logging context.

For this first revision, I've included hooks to (a) annotate each log
message with the logging call site, and (b) include stack traces for all
log messages above a given level.