Replace io.Writer with zap.WriteSyncer
We want to flush our output files whenever we log at the panic or fatal
levels, otherwise there's a chance the program will crash before we
actually write our logs. This is particularly important if we use
a buffered output writer.

To accomodate this, replace our use of io.Writer with zap.WriteSyncer.
All *os.Files already satisfy our interface, and this diff adds
a wrapper for other io.Writers.