Add a stdlib- and bark-compatible wrappers (#28)
* Add an UnsafeBytes field type

Add an escape hatch, so that other packages can access the fields'
underlying buffer more directly. This is, unfortunately, a necessary
precursor to building a bark-compatible wrapper.

* Add a bark-compatible wrapper

* Add a stdlib-compatible wrapper

* Add benchmarks for wrapper packages

* Add wrapper packages to Makefile

* Add bark to Glide manifest

* Remove unnecessary bark file

* Don't export unsafe byte methods

Rather than exporting the UnsafeAddBytes methods and fields, have zbark upcast zap.Logger to access a method not included in the Logger interface. This makes zbark gross, but keeps zap nice and clean.