Encode errors under different keys
Currently, we don't handle encoding errors in a consistent way
- `AddObject` serializes the error into the log message, but
`AddMarshaler` logs the message to standard out. Even though very few
portions of the code can fail at runtime, we end up checking and
managing errors all over the place.

This PR allows the encoder to return an error for `AddObject` and
`AddMarshaler`, which are the only two methods that can fail. However,
the field handles that error by serializing it under a different key
(errors for the key `foo` are serialized under `fooError`).

This should make our APIs less cumbersome and our error handling more
consistent. It fixes #43.