Support concurrent logging to a single WriteSyncer
Currently, the WriteSyncer assumes the underlying Writer will support
concurrent Writes correctly. However, this is not true for common
io.Writer implementations:
 - *os.File may call the write syscall multiple times
 - *bytes.Buffer does not support concurrent writes at all

Instead, we will wrap all passed in WriteSyncers in a lockedWriteSyncer
which will use a mutex to serialize calls to the underlying WriteSyncer.

In future, we may want to expose a way of setting a WriteSyncer that
already supports concurrent calls, but since most common implementations
cannot handle concurrent writes correctly, locking the writer seems like
a simpler default.