Add options for JSON encoders (#105)
* Add JSON options to format message, time, and level

Add options for the JSON encoder, and supply options to control the
message, timestamp, and level formatting. Users can also supply their
own formatters if they have unusual requirements.

* Support formatting options in JSON encoder

Allow users to pass options in the JSON encoder constructor, and
validate that those options are honored.

To minimize the test changes required, introduce a performance
regression in `WriteEntry`. When we flatten the serialized message (in
the next PR), we'll remove the extra truncate-and-copy step.

* Address CR feedback

Unexport some options-related functionality, remove some stubs, and
clean up multiple frees.