Flatten serialized entries
Log aggregation and processing systems typically require some
information in the top-level JSON. In zap, we included the message,
level, and timestamp, but that's clearly not enough to support many use
cases. We could support adding information to this envelope via a
separate API, but the simplest option is to flatten the entry.

This is a small change to the production code, but it requires large
changes to the tests. To keep review semi-sane, this commit includes
only the production code changes.