Fix Check not triggering panic/exit on Panic/Fatal
Panic and Fatal should always cause a panic or exit, even if the level
is disabled. This is handled correctly in logger.Panic/Fatal/Log, but
when the CheckedMessage is used, we're not calling panic/exit.

We should always return a valid message from Check when the level is
Fatal or Panic. This means that Panic/Fatal will not be sampled, since
the sampler should not call panic/exit directly, but should defer to the
underlying logger, and there's no way to specify don't log, but call the
panic/exit functions.