Reduce allocations in EntryCaller.String
Since we know the types of the file name and line number, we don't need to use fmt.Sprintf and pay that allocation cost. We can instead manage a byte slice ourselves, which saves a few small allocations.

(Eventually, we should pool these byte buffers to save a further allocation.)