Don't expose structured fields or checking
Keep the sugared API wholly separate from the unsugared API: don't expose a way
to add plan `zapcore.Field`s to the context, and don't plan to expose a checked
entry API. This cleans up the sugared logger implementation.

Since the sugared logger is a concrete type and not an interface, we can always
add these curlicues back later.