Double down on loosely-typed fields for SugaredLogger (#251)
* Double down on loosely-typed fields for SugaredLogger

Some of zap's strongly-typed fields rely on the field constructors to do some
work, so it's difficult for users to replicate their effects using the variadic
key-value pairs that the sugared logger prefers. (Stack() is a good example of
this.)

To accommodate these use cases, the sugared logger has a `WithFields` method,
but this isn't very satisfying; for symmetry, we'd also want `InfoWithFields`,
`DebugWithFields`, etc.

This PR doubles down on developer ergonomics (over compile-time safety) by
making `With`, `Debugw`, `Infow`, etc. able to accept a mix of zapcore.Fields
and loosely-typed key-value pairs.

* Clean up sweetenFields state machine and test

* Log all non-string keys together

* Update comment on With

* Add some more test coverage