Add encoder support for most built-in types (#258)
* Add encoder support for most built-in types

iAs part of finalizing the encoder interface, this PR adds methods to support
nearly all the built-in types. This greatly expands the interface, but it's
desirable for binary encodings (like msgpack) that will want to treat each size
of integer differently. Since we're adding so many methods to the encoders
anyways, I rounded things out by supporting all the built-in types except error.

Overall, I think that this is a good tradeoff - I expect that there will be many
Facility and WriteSyncer implementations in the wild, but relatively few
Encoders.

* Replace FIXME