Remove Nest field constructor
In a year of using zap, we haven't needed this constructor. It's also going to
be a bit confusing to explain the differences between `Nest` and the forthcoming
`Scope`, so let's just drop `Nest` until we're sure we need it. We can always
add it post-1.0, and users can still use the `Fields` type directly.