Add support for logging to isolated namespaces (#265)
* Remove Nest field constructor

In a year of using zap, we haven't needed this constructor. It's also going to
be a bit confusing to explain the differences between `Nest` and the forthcoming
`Scope`, so let's just drop `Nest` until we're sure we need it. We can always
add it post-1.0, and users can still use the `Fields` type directly.

* Alter map-based encoder to support namespaces

In order to support the isolated namespaces proposed in #132, `MapObjectEncoder`
needs to maintain more internal state. This requires converting it to a struct
instead of a simple `map[string]interface{}`.

* Add namespace support to encoders and loggers

Add an `OpenNamespace` method to encoders and create a `NamespaceType` field.
This allows applications to force sub-components and third-party libraries to
log into an isolated namespace, preventing key collisions.

Fixes #132.

* Fixes from CR