Export Level-, Time-, and DurationEncoders (#279)
* Export Level-, Time-, and DurationEncoders

This is another precursor to terse, config-based logger construction. Since we'd
like users to be able to hydrate a logger from JSON or YAML config, we must
provide a way for them to specify how they'd like levels, times, and durations
to be encoded. We'll do so by exporting a small selection of encoders and
implementing text.Unmarshaler.

Note that this changes the default time encoding to floating-point seconds since
epoch, which is IMO a better default than our current integer
milliseconds-since-epoch. I plan to keep this the default for non-development
loggers.

* Make SecondsDurationEncoder the default