Rewrite sampler to fix oversampling
See discussion on how the current sampler is wrong:
https://github.com/uber-go/zap/pull/252#discussion_r99447146

Instead of using `time.AfterFunc` and resetting timers, use timestamps
on the entries, and maintain a sliding window for the current "tick"
with atomics to track when the counters should be reset.

The alternate approach is to start the timer as soon as the first
message for any bucket is logged, but that would cause more timers to be
started, which costs an allocation as well as requiring a global lock on
the timer heap.