Fix allocation when returning []byte to pool (#289)
In order to pool byte slices without incurring an allocation (by casting into
interface{}), we need to switch to using a pointer type. The natural candidate
is the standard library's bytes.Buffer; it's a bit slower than straight
slices, but that's probably okay. However, we use strconv.AppendFloat and
friends to avoid allocations in the encoder, and those functions won't work with
byte.Buffer.

Instead, this PR adds a small Buffer type that implements just the bits of
bytes.Buffer and strconv that we need. It renames the private buffers
package to bufferpool to clarify the difference between the two.