Rename Facility to Core (#291)
Rename zapcore.Facility to zapcore.Core, which is a bit more evocative.
zapcore.Logger is probably a better name, but is likely to create more confusion
among first-time users and GoDoc readers.

After looking at the end result, I'm inclined to stop here and keep Logger and
SugaredLogger as they are. Elevating SugaredLogger to the primary type ended up
(to me) confusing the story of what's the fundamental logger and what's the
API sugaring layer.