Support non-standard, rich error types (#325)
Users who take the time (and performance hit) of using Dave Cheney's errors
package should get rich, annotated errors and stacktraces in their logs.
However, that package is only one of a number of similar packages and hasn't yet
committed to a stable API, so we don't want to take a direct dependency on it.

To support these sorts of packages, serialize errors that implement
fmt.Formatter differently from those that don't. This is necessarily slower than
serializing standard errors, but users of these packages have already committed
to a more expensive error-handling path.