Add ByteString API (#366)
This is a breaking change that adds bytestring APIs (to log UTF-8 encoded bytes) to ObjectEncoder and ArrayEncoder. To actually save allocations along this path, I also benchmarked adding a []byte to the field union; this reduces the allocation count for
bytestrings, but increases the cost of adding fields by ~50%.

Fixes #324.