zapcore.Level: Move MarshalText to value receiver
Currently, MarshalText and UnmarshalText are both using pointer
receivers but only UnmarshalText needs a pointer receiver since it
mutates the underlying value.

MarshalText should be on the same receiver type as String. On the other
hand, UnmarshalText requires a pointer receiver, and it can handle the
case where we try to unmarshal to a nil pointer.