Move error logging code to error.go
This gives errors their own file like array.go before we add more
complex handling for error types.

I also moved the tests for errors to error_test to be consistent.

Feel free to close this if you don't want to organize the code this way,
in which case the multierr integration will be in field.go.