Add documentation and examples for zap
Add a bunch of testable examples and documentation for the main `zap`
package. As we get closer to launching fx and pushing zap adoption
internally, documentation will get more and more important.

We'll still need to improve the documentation of the other packages, but
this addresses the lion's share of #139.