Remove unused code to ignore runtime stack frames
runtime.CallerFrames is currently ignoring the last frame, which is
always runtime.main (for the main goroutine) and runtime.goexit (for
additional goroutines). Since the main goroutine will include the user's
main function, this is noise that doesn't add much value.

We had special logic to ignore these runtime frames specifically, but
it's not required since it's only ever the very last stack frame and we
already ignore the last frame.