Make zaptest a wrapper around internal/ztest
Mitigate the distasteful copy-paste of `zaptest` by providing a path
forward using Go 1.9 aliases. For Go 1.9 users, `zaptest` is now nothing
more than a few type aliases; this cleans up the GoDoc output nicely,
since it removes the documentation for all the methods on the testing
spies while remaining fully backward-compatible. For Go 1.8 and earlier,
we keep the original code but add a test to make sure that it exactly
matches the version in `internal/ztest`.

With this change, we're free to introduce functionality into `zaptest`
that relies on `zapcore` and `zap` itself. In particular, we can return
`*zap.Logger` from constructors.