Add table of extensions to FAQ
As part of the discussion on https://github.com/uber-go/zap/issues/96,
it's clear that we need to make it easier for third-party packages to
extend zap's functionality. Some of that requires new APIs, but we'll
also need a way to refer people from the core zap package to the various
extensions.

This PR adds an FAQ entry that lists the extensions we're aware of. It's
currently a little hard to find (the GoDoc includes a link to the FAQ,
but it's not very prominent). We'll address that as part of a
documentation revamp later on.