Add an alias for zapcore.Field
The `Field` union struct is defined in `zapcore`, but most of its
constructors are in the top-level `zap` package. This makes the zap
GoDoc really unfriendly, since the field constructors are mixed in with
other top-level functions instead of being grouped together under the
field type.

This commit adds a `zap.Field` type which is just an alias for
`zapcore.Field`. It then changes all the top-level zap code to use the
alias. For validation, this commit leaves all test code unchanged.