Re-run benchmarks with the latest dependencies
Benchmarks were last run 2y ago, so it's time to see how the
lay of the land has changed.

Just baseline wise it appears that my laptop is about 50%
faster than the hardware that the initial benchmarks were
ran on, but it doesn't really change the relative difference
in the loggers.

Zerolog appears to have closed the distance quite a bit,
while others are still an order of magnitude slower than
zap.