license: allow github to detect MIT license
Resolves #731

github uses https://github.com/licensee/licensee to detect
the license for the project.  Because the project contains a file
named "check_license.sh" it thinks this is a license file.  It is
unable to detect a license for that file and fails.  Here is the
run output:

< ./bin/licensee detect ../zap/
License:        NOASSERTION
Matched files:  LICENSE.txt, check_license.sh
LICENSE.txt:
  Content hash:  d64f3bb4282a97b37454b5bb96a8a264a3363dc3
  Attribution:   Copyright (c) 2016-2017 Uber Technologies, Inc.
  Confidence:    100.00%
  Matcher:       Licensee::Matchers::Exact
  License:       MIT
check_license.sh:
  Content hash:  2ca67b08b61b91f2b3c67fe4581ee9e8dc9da2e8
  License:       NOASSERTION

If you rename the file to checklicense.sh it won't think that it is a
license file and will correctly advertise zap as a MIT project

< ./bin/licensee detect ../zap/
License:        MIT
Matched files:  LICENSE.txt
LICENSE.txt:
  Content hash:  d64f3bb4282a97b37454b5bb96a8a264a3363dc3
  Attribution:   Copyright (c) 2016-2017 Uber Technologies, Inc.
  Confidence:    100.00%
  Matcher:       Licensee::Matchers::Exact
  License:       MIT