TestStacktraceFiltersVendorZap: Fix for modules
This fixes TestStacktraceFiltersVendorZap to work with Go modules.

For context, the test previously did the following:

- Set up a temporary GOPATH
- Copy stacktrace_ext_test.go into it
- Symlink Zap and relevant dependencies into a vendor/ directory in the
  temporary GOPATH
- Run specific tests from stacktrace_ext_test.go with Zap inside the
  vendor/ directory

We're no longer using GOPATH or vendor/ directories, but this behavior
of Zap when *it* is inside a vendor/ directory should be retained.

To that end, this changes TestStacktraceFiltersVendorZap to continue to
run with the following changes:

- Use `go mod download -json` to retrieve the locations of all of Zap's
  dependencies on-disk, and symlink them into the temporary GOPATH's
  vendor/.
- Explicitly set `GO111MODULE=off` before running `go test` because we
  may be running in an environment where `GO111MODULE=on` is set.