ci: Use Go modules and ./...
This switches the Makefile and travis setup to rely on Go modules and
`./...` instead of `glide nv`.

Note that a tradeoff of making `benchmarks/` its own module is that
`./...` in the Zap directory will no longer include the `benchmarks/`
directory. So where necessary, we need to run commands separately in
that directory as well, which complicates the build script ever so
slightly.