json: Don't panic for nil Encode{Time, Duration}
Fixes #834

The JSON encoder assumes that encoders for `time.Time` and
`time.Duration` are always specified, which causes nil pointer
dereference panics.

Fix this by treating nil encoders for time and duration as no-ops. This
will fall back to existing logic in the JSON encoder that handles no-op
time and duration encoders.