${key}Verbose and ${key}Causes for wrapped errors
encodeError cannot provide ${key}Verbose and ${key}Causes fields
if the error implements fmt.Formatter and/or errorGroup interfaces
but is wrapped.

This change uses errors.As instead of a type switch to check if
the given error and the underlying errors are compatible with the
above interfaces, and computes ${key}Verbose and ${key}Causes fields
with unwrapped error.

This commit also upgrades github.com/pkg/errors to v0.9.1 because
v0.8.1 doesn't support Unwrap method.