fossa: Run separately, only on push
Currently, the FOSSA analysis is set to run as part of CI. Minus the
fact that it's not really part of the build, its reliance on a secret
means that it won't run for any pull requests made from external forks.

Resolve this by running the FOSSA analysis only when we push to a
branch of the Zap repository.