JSON encoder: Use correct precision for Complex64
Converting to Complex128 meant that the real and imaginary components of
the Complex64 were encoded as float64s which, like #1002, ended up
those components with the incorrect precision.

Fix this by adding a precision parameter to appendComplex tp specify the
precision with which the two components of the complex number should be
encoded.