Drop support for Go < 1.12
Go 1.12 is old enough that we don't need to try to support it.
The files global_go112 and global_prego112 existed to
support different stack depths for the standard library logger,
which changed going from Go 1.11 to 1.12.

This is no longer something we need to support because
per our CI configuration, we only support Go 1.15+.

Ref #682