CI: Run against Go 1.17 and 1.18
We support the two most recent Go minor relases with Zap.
With 1.18 newly released, we can transition to only 1.17 and 1.18.

I took this opportunity to also bump the Go version specified in the
go.mod files since there is a change in behavior between when we
generated these and now.

staticcheck had to be upgraded to master to run successfully.