Add ObjectValues field constructor
In the previous change, we added a Objects field constructor to support
logging arrays of any object supported by zap.Object.

    func (*User) MarshalLogObject(...) error { ... }
    var users []*User = ...
    zap.Object("users", users)

That solution is incomplete because it does not cover the case when,
instead of a `[]*User`, we have a `[]User` despite the method being on
`*User`.

This change adds a similar ObjectValues field constructor that handles the
case where we have a `[]T` but the MarshalLogObject is on `*T`.

This works by declaring a new type constraint objectMarshalerPtr.
`objectMarshalerPtr[T]` is true for a type that is a pointer to `T`,
and implements zapcore.ObjectMarshaler.

Note that we still do not cover the case where the method is on the value
receiver (`func (User) MarshalLogObject(..) error`) and we have a
`[]*User`. We can add that when it becomes necessary.

Reviewed-At: D7295739
Reviewed-By: ztstewart