Refactor to reuse Log method
This refactors the implementation of the Log method to reuse it
in the other level variants of the method (i.e. `Debug`, `Info`, etc.).

To do this, a skip argument was added to `check` to allow its callers to
specify how many frames to skip in the log, and a private `log` method
was added which calls this `check` method with the appropriate frame
skip count.