Open: expose error cause, test for fs.ErrNotExist (#1149)
This is a prefactor for Windows support (#621).
TestOpen currently relies on a hardcoded error "no such file or
directory" if a file is not found. However, this error message is
OS-specific.  We can now rely on `errors.Is(err, fs.ErrNotExist)`
if we wrap errors using %w instead of %v.

Co-authored-by: Abhinav Gupta <mail@abhinavg.net>