ci: Set up dependabot for GH actions and tools
Sets up dependabot to update GitHub Actions
so that we're not running severely outdated actions (#1235)
and catching up in one go.

Similarly, sets up dependabot for go.mod updates
of the tools submodule so we get the latest staticcheck.

Note that this does NOT add dependabot for Zap itself.
The reasoning for this is that Zap is a library.
If we keep require latest versions of all of Zap's dependencies,
then users of Zap will be forced to upgrade to latest versions
of all transitive dependencies, even if they don't want to/can't.